def struct
